core: Separate better Context and State structs
