+++ core/src/subgraph/mod.rs
@@ -1,3 +1,4 @@
+mod context;
 mod error;
 mod inputs;
 mod instance;
